requirements.txt
# Python Requirements | Cannlytics API
# Created 2/1/2021

# To ensure app dependencies are ported from your
# virtual environment/host machine into your container,
# run 'pip freeze > requirements.txt' in the terminal to overwrite this file.

datetime==4.3
django==3.0.7
django-environ==0.4.5
djangorestframework==3.12.2
firebase_admin==4.3.0
google-auth==1.20.1
google-cloud-secret-manager==1.0.0
gunicorn==20.0.4
pandas==1.2.1
python-decouple==3.3
psycopg2-binary==2.8.5
